% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCloopFunction.R
\name{ABCloop}
\alias{ABCloop}
\title{Approximate Bayesian Calibration Algorithm}
\usage{
ABCloop(
  N0 = 1000,
  N.centers = 1,
  B = 10,
  N.post = 1,
  max.iter = 1,
  seed = 1234,
  N.cov.points = 0,
  sample.inflate = 1,
  inputs.dir = "calib_inputs",
  prior.file = "parmpriors.csv",
  default.file = "default_parm_values.csv",
  target.sims = "targetspecs.csv",
  target.file = "calibrationtargets.xlsx",
  model.parm.order = "model_parm_order.csv",
  userfn.dir = "calib_userfn",
  outfile.modelparms = "modelparms.csv",
  outfile.dist = "distance.csv",
  outfile.simparms = "simparms.csv",
  outfile.sampling = "meancov.csv",
  output.directory = ".",
  LHS = TRUE,
  continue.runs = FALSE,
  reeval = FALSE,
  prevruns.dir = ".",
  print.status = FALSE,
  recalc.centers = TRUE
)
}
\arguments{
\item{N0}{initial draw size}

\item{N.centers}{initial centers}

\item{B}{PLACEHOLDER}

\item{N.post}{PLACEHOLDER}

\item{max.iter}{PLACEHOLDER}

\item{seed}{overall seed value, used to simulate multiple streams}

\item{N.cov.points}{PLACEHOLDER}

\item{sample.inflate}{PLACEHOLDER}

\item{inputs.dir}{directory containing model inputs}

\item{prior.file}{contains prior distribution information for model parameters}

\item{default.file}{PLACEHOLDER}

\item{target.sims}{specification for simulations for each target}

\item{target.file}{PLACEHOLDER}

\item{model.parm.order}{PLACEHOLDER}

\item{userfn.dir}{dirctory containing user-specified functions}

\item{outfile.modelparms}{PLACEHOLDER}

\item{outfile.dist}{PLACEHOLDER}

\item{outfile.simparms}{PLACEHOLDER}

\item{outfile.sampling}{PLACEHOLDER}

\item{output.directory}{PLACEHOLDER}

\item{LHS}{PLACEHOLDER}

\item{continue.runs}{PLACEHOLDER}

\item{reeval}{reevaluate currently accepted runs (logical)}

\item{prevruns.dir}{PLACEHOLDER}

\item{print.status}{PLACEHOLDER}

\item{recalc.centers}{PLACEHOLDER}
}
\value{
ABCLoopFunction creates 7 output files, results.dir is the results directory:
             1. outfile.modelparms: all drawn model parameters
             2. outfile.dist: distance from targets
             3. outfile.simparms: simulated targets at model parameters
             4. 'good.'outfile.modelparms: accepted model parameters
             5. 'good.'outfile.dist: distance from targets for accepted model parameters
             6. 'good.'outfile.simparms: simulated targets at acccepted model parameters
             7. outfile.sampling: mean and covariance of sampling distributions
}
\description{
ABCLoopFunction creates 7 output files, results.dir is the results directory:
             1. outfile.modelparms: all drawn model parameters
             2. outfile.dist: distance from targets
             3. outfile.simparms: simulated targets at model parameters
             4. 'good.'outfile.modelparms: accepted model parameters
             5. 'good.'outfile.dist: distance from targets for accepted model parameters
             6. 'good.'outfile.simparms: simulated targets at acccepted model parameters
             7. outfile.sampling: mean and covariance of sampling distributions
}
\note{
While biased targets can be included, and are used for the Corley targets, the current implementation assumes
that estimtaes are downwardly biased. Code would need ot be generalized to allow upward bias.

We incorporate one tpe of user-specified prior distribution: a resticted uniform, which is used to accomodate the
restricted range of beta2 growth parameters, which depend on beta1

There is currently some specialized coding to allow continuation of prior runs. Continuation requres the use of the
same targets. Some code that can be removed correcs an issue with naming of SEER targets
}
