% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prior.R
\name{add_prior}
\alias{add_prior}
\title{Add Prior Distribution Information}
\usage{
add_prior(
  ...,
  dist_base_name = NULL,
  density_fn = NULL,
  quantile_fn = NULL,
  parameter_name = NULL
)
}
\arguments{
\item{...}{Parameters read by prior generating function}

\item{quantile_fn}{}
}
\value{

}
\description{
Add Prior Distribution Information
}
\details{
If the user doesn't provide any RNG functions, imabc will create a parameter that doesn't vary (in the future).
If the user only provides one of the required RNG functions, add_prior will search for the most logical name of the other
function. I.e. if a density function is provided (such as dunif), add_prior will look for the quantile function (qunif)
If the user selects both functions independently or create their own functions it is on them to make sure they work
appropriately.
}
