% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_target.R
\name{add_target}
\alias{add_target}
\title{Title}
\usage{
add_target(
  target,
  starting_range,
  stopping_range,
  FUN = NULL,
  target_name = NULL
)
}
\arguments{
\item{target_name}{}
}
\value{

}
\description{
Title
}
