% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.alpha.R
\name{calc.alpha}
\alias{calc.alpha}
\title{Title}
\usage{
calc.alpha(sim.p, sim.p.names, target.data, specs = NULL)
}
\arguments{
\item{sim.p}{PLACEHOLDER}

\item{sim.p.names}{PLACEHOLDER}

\item{target.data}{PLACEHOLDER}

\item{specs}{PLACEHOLDER}
}
\value{

}
\description{
Title
}
