% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_run_dt.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{Initialize Run data.tables}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Initialize Run data.tables
}
\keyword{datasets}
