% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello()
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
You can learn more about package authoring with RStudio at:

  http://r-pkgs.had.co.nz/

Some useful keyboard shortcuts for package authoring:

  Install Package:            'Cmd + Shift + B'
  Check Package:              'Cmd + Shift + E'
  Test Package:               'Cmd + Shift + T'
  Generate roxygen2 temp:     'Cmd + Shift + Option + R'
  load functions pre-install: devtools::load_all()
  Build roxygen2 docs:        devtools::document()
  Build package:              devtools::build()
}
