% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_previous_results.R
\name{read_previous_results}
\alias{read_previous_results}
\title{Read Previous Results}
\usage{
read_previous_results(path, tag = NULL)
}
\arguments{
\item{path}{character(1). The location of files saved during a previous run.}

\item{tag}{Optional character(1). If multiple runs have been saved to a single path, provide the tag that differentiates them.}
}
\description{
Searches the files found in path for the files saved by an imabc run and reads them into the current environment.
}
\note{
tag is required if multiple sets of results are stored in a single location.
}
