% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imabc.R
\name{imabc}
\alias{imabc}
\title{imabc}
\usage{
imabc(
  priors,
  targets,
  target_fun,
  previous_results = NULL,
  N_start = 1,
  seed = 12345,
  latinHypercube = TRUE,
  N_centers = 1,
  Center_n = 50,
  N_post = 100,
  max_iter = 1000,
  N_cov_points = 0,
  sample_inflate = 1,
  recalc_centers = TRUE,
  backend_fun = NULL,
  output_directory = NULL,
  output_tag = "timestamp",
  verbose = TRUE
)
}
\arguments{
\item{priors}{list. A list with all the information required by define_priors and add_priors}

\item{targets}{list. A list of the main targets and their subtargets}

\item{target_fun}{function. A function that takes in parameters and returns the predicted subtarget values}

\item{previous_results}{list. optional.}

\item{latinHypercube}{Only needed if previous_results not included}

\item{verbose}{}
}
\value{

}
\description{
imabc
}
