% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imabc.R
\name{imabc}
\alias{imabc}
\title{Incremental Mixture Approximate Bayesian Computation (IMABC)}
\usage{
imabc(
  priors,
  targets,
  target_fun,
  previous_results = NULL,
  N_start = 1,
  seed = NULL,
  latinHypercube = TRUE,
  N_centers = 1,
  Center_n = 50,
  N_post = 100,
  max_iter = 1000,
  N_cov_points = 0,
  sample_inflate = 1,
  backend_fun = NULL,
  output_directory = NULL,
  output_tag = "timestamp",
  verbose = TRUE,
  validate_run = FALSE
)
}
\arguments{
\item{priors}{A priors object created using define_priors. This contains information regarding the parameters that
are being calibrated}

\item{targets}{A targets object created using define_targets. This contains information regarding the target values
which will be used to evaluate simulated parameters.}

\item{target_fun}{A function, that will tranform parameters into target values. To ensure that the function takes in
the correct values and returns the results correctly it is strongly advised to use define_target_function() for this.}

\item{previous_results}{Optional list of results from a previous}

\item{latinHypercube}{Only needed if previous_results not included}

\item{verbose}{}
}
\value{

}
\description{
Calibrates a set of parameters following the IMABC method.
}
